Fix build errors and adjust auth setup
