+++ src/app/api/bookings/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
 import { prisma } from '@/lib/prisma';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../auth/[...nextauth]/route';
+import { authOptions } from '@/lib/auth';
 
 // ── GET /api/bookings?userId=…&status=… ───────────────────────────────────────
 export async function GET(request: Request) {
