+++ src/app/api/customer/bookings/route.ts
@@ -1,7 +1,7 @@
 // File: src/app/api/customer/bookings/route.ts
 import { NextRequest, NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../../auth/[...nextauth]/route';
+import { authOptions } from '@/lib/auth';
 import { prisma } from '@/lib/prisma';
 
 // POST /api/customer/bookings
