+++ src/types/react-beautiful-dnd.d.ts
@@ -1,0 +1,1 @@
+declare module 'react-beautiful-dnd';
